requirements.txt
Django==4.0.5
mysqlclient==2.1.0
Pillow==9.1.1